all 3 files in the project root has been moved to a folder called generators
