+++ generators/generate_carousel.py
