+++ generators/generate_quotes.py
